requirements.txt
sounddevice
numpy
scipy
openwakeword
onnxruntime
ollama
duckduckgo-search
Pillow
ddgs
faster-whisper
demoji
piper-tts